pyproject.toml
[tool.poetry]
name = "repl_python3_DiligentTestyOpengl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
rsa = "^4.8"
pycrypto = "^2.6.1"
colorama = "^0.4.4"
network = "^0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
